requirements.txt
jsonpickle
python-magic
python-gnupg>=0.5
py3-validate-email; python_version >= '3.8'
py3-validate-email<1.0.6; python_version < '3.8'
